rename file appropriately
